requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coverage==7.8.0
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.12
firecrawl==2.1.2
frozenlist==1.6.0
griffe==1.7.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.9.0
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.4.3
nest-asyncio==1.6.0
openai==1.72.0
openai-agents==0.0.12
packaging==25.0
pluggy==1.5.0
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.46.1
tomli==2.2.1
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.28.0
websockets==15.0.1
yarl==1.20.0
aioresponses==0.7.8
logfire[fastapi]==3.20.0
